-exported types, refactored typings
