+++ src/index.ts
@@ -126,3 +126,4 @@ export { default as dialogBehavior } from './lib/accessibility/Behaviors/Dialog/
 //
 export { default as mergeThemes } from './lib/mergeThemes'
 export { createComponent } from './lib'
+export { RenderStardustResultConfig, CreateStardustComponentConfig } from '../types/utils'
