+++ types/utils.d.ts
@@ -3,6 +3,7 @@
 // ========================================================
 
 import * as React from 'react'
+import { ComponentSlotClasses } from '../src/themes/types'
 
 export type Extendable<T> = T & {
   [key: string]: any
@@ -17,6 +18,16 @@ export type OneOrArray<T> = T | T[]
 export type ObjectOf<T> = { [key: string]: T }
 
 export type ObjectOrFunc<TResult, TArg = {}> = ((arg: TArg) => TResult) | TResult
+
+export interface RenderStardustResultConfig {
+  classes: ComponentSlotClasses
+  rtl: boolean
+}
+
+export interface CreateStardustComponentConfig<P> {
+  displayName?: string
+  render: (config: RenderStardustResultConfig, props: P) => React.ReactNode
+}
 
 // ========================================================
 // Props
